renamed PygrationSet to be Loader
this is step one in making pygrate fully testable.  new name better
describes that the clas does and makes room for a PygrationSet that
only contains objects rather than also loading them from the filesystem

it shouldn't break
the code b/c the class is the same just with a different name.  if it
does break, will clean it up later.