created tests for migrator class
migrator wasn't tested, that's bad.  to do this, needed to move the session
out of the migrator and into the LiveDB class.  that also moved committing
into the DB class where it needed to be anyway to avoid committing in test
or Noop mode.  Also changed 'pass' state flag to just 'P'